mcp_server/requirements.txt
fastapi
uvicorn
pyyaml
requests
python-dotenv
